% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absent.R
\name{absent}
\alias{absent}
\title{Check if an activity is absent from a case}
\usage{
absent(activity, n = 1)
}
\arguments{
\item{activity}{The activity too check}

\item{n}{n-1 is the allowed number of occurences of the activity. E.g. n = 1 means the activity should be absent, n = 2 means it is allowed to occur once.}
}
\description{
Check if an activity is absent from a case
}
