% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_exists.R
\name{co_exists}
\alias{co_exists}
\title{Check for co-existence of two activities}
\usage{
co_exists(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A exists, Activity B should also exist, and vice versa.
}
