% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ends.R
\name{ends}
\alias{ends}
\title{Check if cases end with an activity}
\usage{
ends(activity)
}
\arguments{
\item{activity}{The end activity. Character vector of length one.}
}
\description{
Check if cases end with an activity
}
