% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{exists}
\alias{exists}
\title{Check if activity is present (exists) in a case}
\usage{
exists(activity, n = 1)
}
\arguments{
\item{activity}{Activity to check}

\item{n}{The minimum number of times the activity should be present.}
}
\description{
Check if activity is present (exists) in a case
}
