% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_between.R
\name{exists_between}
\alias{exists_between}
\title{Check if activity is present (exists) in a case between  min and max number of times}
\usage{
exists_between(activity, min = 1, max = 1)
}
\arguments{
\item{activity}{Activity too check}

\item{min}{The minimum number of times the activity should be present.}

\item{max}{The maximum number of times the activity should be present.}
}
\description{
Check if activity is present (exists) in a case between  min and max number of times
}
