% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_exactly.R
\name{exists_exactly}
\alias{exists_exactly}
\title{Check if activity is present (exists) in a case for exactly n times}
\usage{
exists_exactly(activity, n = 1)
}
\arguments{
\item{activity}{Activity too check}

\item{n}{The exact number of times the activity should be present.}
}
\description{
Check if activity is present (exists) in a case for exactly n times
}
