% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responded_existence.R
\name{responded_existence}
\alias{responded_existence}
\title{Check for responded existence between two activity}
\usage{
responded_existence(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A occurs in a case, activity B should also occur (before or after).
}
