% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Response}
\usage{
response(activity_a, activity_b)
}
\arguments{
\item{activity_a}{\code{\link{character}}: Activity A. This should be an activity of the log supplied to \code{\link{check_rule}}.}

\item{activity_b}{\code{\link{character}}: Activity B. This should be an activity of the log supplied to \code{\link{check_rule}}.}
}
\description{
Check for response between two activities.

If \code{activity_a} is executed, it should be (eventually) followed by \code{activity_b}.
}
\examples{
library(bupaR)
library(eventdataR)

# A blood test should eventually be followed by Discuss Results.
patients \%>\%
 check_rule(response("Blood test","Discuss Results"))

}
\seealso{
\code{\link{precedence}},\code{\link{responded_existence}},\code{\link{succession}}

Other Declarative Rules: 
\code{\link{absent}()},
\code{\link{and}()},
\code{\link{contains_between}()},
\code{\link{contains_exactly}()},
\code{\link{contains}()},
\code{\link{ends}()},
\code{\link{precedence}()},
\code{\link{responded_existence}()},
\code{\link{starts}()},
\code{\link{succession}()},
\code{\link{xor}()}
}
\concept{Declarative Rules}
