% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{and}
\alias{and}
\title{Check for co-existence of two activities}
\usage{
and(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A exists, Activity B should also exist, and vice versa.
}
