% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rule.R
\name{check_rule}
\alias{check_rule}
\title{Check declarative rules.}
\usage{
check_rule(eventlog, rule, label = NULL)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{rule}{A rule create by a rule function.}

\item{label}{Optionally, the variable name under which the result of the rule should be stored.}
}
\description{
Check declarative rules.
}
