% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rules.R
\name{check_rules}
\alias{check_rules}
\alias{check_rules.log}
\title{Check Multiple Declarative Rules}
\usage{
check_rules(log, ..., eventlog = deprecated())

\method{check_rules}{log}(log, ..., eventlog = deprecated())
}
\arguments{
\item{log}{\code{\link[bupaR]{log}}: Object of class \code{\link[bupaR]{log}} or derivatives (\code{\link[bupaR]{grouped_log}}, \code{\link[bupaR]{eventlog}}, \code{\link[bupaR]{activitylog}}, etc.).}

\item{...}{Name-rule pairs created by rule functions.}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\value{
An annotated log (of same type as input), where – for every rule – a new column indicates whether the rule holds or not.
The name of each rule becomes the name of the column.
}
\description{
This function can be used to check rules on event data. It needs a \code{log} (object of class \code{\link[bupaR]{log}} or derivatives,
e.g. \code{\link[bupaR]{grouped_log}}, \code{\link[bupaR]{eventlog}}, \code{\link[bupaR]{activitylog}}, etc.). and a \code{rule}.
Rules can be made with the following functions:
\itemize{
\item \code{\link{absent}},
\item \code{\link{and}},
\item \code{\link{contains}},
\item \code{\link{contains_between}},
\item \code{\link{contains_exactly}},
\item \code{\link{ends}},
\item \code{\link{precedence}},
\item \code{\link{response}},
\item \code{\link{responded_existence}},
\item \code{\link{starts}},
\item \code{\link{succession}},
\item \code{\link{xor}}.
}
}
\details{
When \code{check_rules} is applied to a \code{\link[bupaR]{grouped_log}}, the grouping variables are ignored but retained in the returned log.
}
\section{Methods (by class)}{
\itemize{
\item \code{check_rules(log)}: Check rules on a \code{\link[bupaR]{log}}.

}}
\examples{
library(bupaR)
library(eventdataR)

# Check whether MRI Scan is preceded by Blood test, and the case starts with Registration.
patients \%>\%
 check_rules(rule1 = precedence("Blood test","MRI SCAN"),
             rule2 = starts("Registration"))

}
\seealso{
\code{\link{check_rule}}
}
