% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ends.R
\name{ends}
\alias{ends}
\title{Ends}
\usage{
ends(activity)
}
\arguments{
\item{activity}{\code{\link{character}}: The end activity. This should be an activity of the log supplied to \code{\link{check_rule}}.}
}
\description{
Check if cases end with the specified activity.
}
\examples{
library(bupaR)
library(eventdataR)

# A patient's last activity should be the Check-out
patients \%>\%
 check_rule(ends("Check-out"))

}
\seealso{
Other Existence rules: 
\code{\link{absent}()},
\code{\link{contains_between}()},
\code{\link{contains_exactly}()},
\code{\link{contains}()},
\code{\link{starts}()}
}
\concept{Existence rules}
