% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence.R
\name{precedence}
\alias{precedence}
\title{Precedence}
\usage{
precedence(activity_a, activity_b)
}
\arguments{
\item{activity_a}{\code{\link{character}}: Activity A. This should be an activity of the log supplied to \code{\link{check_rule}}.}

\item{activity_b}{\code{\link{character}}: Activity B. This should be an activity of the log supplied to \code{\link{check_rule}}.}
}
\description{
Check for precedence between two activities.

If \code{activity_b} occured, it should be preceded by \code{activity_a} in the same case.
}
\examples{
library(bupaR)
library(eventdataR)

# A MRI Scan should be preceeded by a Blood test.

patients \%>\%
 check_rule(precedence("Blood test","MRI SCAN"))

}
\seealso{
\code{\link[edeaR]{filter_precedence}},\code{\link{responded_existence}},\code{\link{response}},\code{\link{succession}}

Other Relation rules: 
\code{\link{and}()},
\code{\link{responded_existence}()},
\code{\link{response}()},
\code{\link{succession}()},
\code{\link{xor}()}
}
\concept{Relation rules}
