% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence.R
\name{precedence}
\alias{precedence}
\title{Check for precedence between two activities.}
\usage{
precedence(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity B occured, it should be preceded by activity A
}
