% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Check for response between two activities}
\usage{
response(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A is executed, it should be eventually followed by activity B.
}
