% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts.R
\name{starts}
\alias{starts}
\title{Starts}
\usage{
starts(activity)
}
\arguments{
\item{activity}{\code{\link{character}}: The start activity. This should be an activity of the log supplied to \code{\link{check_rule}}.}
}
\description{
Check if cases start with the specified activity.
}
\examples{
library(bupaR)
library(eventdataR)

# Each patients should first be registered.
patients \%>\%
 check_rule(starts("Registration"))

}
\seealso{
Other Existence rules: 
\code{\link{absent}()},
\code{\link{contains_between}()},
\code{\link{contains_exactly}()},
\code{\link{contains}()},
\code{\link{ends}()}
}
\concept{Existence rules}
