% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts.R
\name{starts}
\alias{starts}
\title{Check if cases start with an activity}
\usage{
starts(activity)
}
\arguments{
\item{activity}{The start activity. Character vector of length one}
}
\description{
Check if cases start with an activity
}
