% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succession.R
\name{succession}
\alias{succession}
\title{Succession}
\usage{
succession(activity_a, activity_b)
}
\arguments{
\item{activity_a}{\code{\link{character}}: Activity A. This should be an activity of the log supplied to \code{\link{check_rule}}.}

\item{activity_b}{\code{\link{character}}: Activity B. This should be an activity of the log supplied to \code{\link{check_rule}}.}
}
\description{
Check for succession between two activities.

If \code{activity_a} happens, it should be eventually followed by \code{activity_b}.
If \code{activity_b} happens, it should be preceded by \code{activity_a}.
}
\examples{
library(bupaR)
library(eventdataR)

# Blood test should always happen before a MRI Scan,
# and both should happen when one of them happens.
patients \%>\%
 check_rule(succession("Blood test","MRI SCAN"))

}
\seealso{
\code{\link{precedence}},\code{\link{responded_existence}},\code{\link{response}}

Other Relation rules: 
\code{\link{and}()},
\code{\link{precedence}()},
\code{\link{responded_existence}()},
\code{\link{response}()},
\code{\link{xor}()}
}
\concept{Relation rules}
