% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succession.R
\name{succession}
\alias{succession}
\title{Check succession between two activities}
\usage{
succession(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A happens, it should be eventually followed by activity B.
If activity B happens, it should be preceded by activity A.
}
