% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor.R
\name{xor}
\alias{xor}
\title{Check for exclusiveness of two activities}
\usage{
xor(activity_a, activity_b)
}
\arguments{
\item{activity_a}{Activity A}

\item{activity_b}{Activity B}
}
\description{
If activity A exists, Activity B should not exist, and vice versa.
}
